requirements.txt
Django==5.1.7
djangorestframework==3.15.2
pandas==2.2.3
scikit-learn==1.6.1
python-dotenv==1.0.1
numpy==2.2.3
python-dateutil==2.9.0.post0
pydantic==2.6.1
